+++ SOD.Common/Plugin.cs
@@ -1,5 +1,6 @@
 ﻿using BepInEx;
 using SOD.Common.BepInEx;
+using System.Reflection;
 
 namespace SOD.Common
 {
@@ -8,7 +9,7 @@ namespace SOD.Common
     {
         public const string PLUGIN_GUID = "Venomaus.SOD.Common";
         public const string PLUGIN_NAME = "SOD.Common";
-        public const string PLUGIN_VERSION = "1.0.3";
+        public const string PLUGIN_VERSION = "1.0.4";
 
         public override void Load()
         {
@@ -17,6 +18,9 @@ namespace SOD.Common
             {
                 Unhollowed_Modules_Folder = System.IO.Path.Combine(Paths.BepInExRootPath, "interop")
             });
+
+            // Apply patches
+            Harmony.PatchAll(Assembly.GetExecutingAssembly());
         }
     }
 }